// src/interfaces/data-firestore.interface.ts
export interface AirNodes {
  airnodes: number;
  network_usage: number;
  timestamp: string;
  users: number;
}

export interface Network {
  t: string;
  network: number;
}

export interface Users {
  t: string;
  users: number;
}
export interface ANdata {
  t: string;
  aNodes: number;
}